hpc_workspace/requirements.txt
# Moondream HPC 依賴套件
# 用法：pip install -r requirements.txt

# 核心模型套件
transformers==4.46.3
einops
accelerate
safetensors

# 圖像處理
Pillow
pyvips

# 工具
tqdm
requests
